+++ src/pipx/commands/common.py
@@ -9,7 +9,7 @@ import userpath  # type: ignore
 from pathlib import Path
 from shutil import which
 from tempfile import TemporaryDirectory
-from typing import Dict, List
+from typing import Collection, Dict, List, Optional
 
 from pipx import constants
 from pipx.colors import bold, red
@@ -112,7 +112,11 @@ def _symlink_package_apps(
 
 
 def get_package_summary(
-    path: Path, *, package: str = None, new_install: bool = False
+    path: Path,
+    *,
+    package: str = None,
+    new_install: bool = False,
+    include_injected: bool = False,
 ) -> str:
     venv = Venv(path)
     python_path = venv.python_path.resolve()
@@ -150,6 +154,7 @@ def get_package_summary(
         new_install,
         exposed_binary_names,
         unavailable_binary_names,
+        venv.pipx_metadata.injected_packages.keys() if include_injected else None,
     )
 
 
@@ -185,6 +190,7 @@ def _get_list_output(
     new_install: bool,
     exposed_binary_names: List[str],
     unavailable_binary_names: List[str],
+    injected_package_names: Optional[Collection[str]] = None,
 ) -> str:
     output = []
     output.append(
@@ -202,6 +208,10 @@ def _get_list_output(
         output.append(
             f"    - {red(name)} (symlink missing or pointing to unexpected location)"
         )
+    if injected_package_names:
+        output.append("    Injected Packages:")
+        for name in injected_package_names:
+            output.append(f"      - {name}")
     return "\n".join(output)